requirements.txt
fastapi==0.103.0
python-dotenv==1.0.0